requirements.txt
attrs==19.1.0
autobahn==19.3.3
backcall==0.1.0
bitstring==3.1.5
bleach==3.1.0
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
Pygments==2.3.1
pyrsistent==0.15.1
pyserial==3.4
python-dateutil==2.8.0
pywinpty==0.5.5
pyzmq==18.0.1
qtconsole==4.4.3
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
txaio==18.8.1
vpnotebook==0.1.3
vpython==7.5.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
win-unicode-console==0.5
